+++ chatbots-telegram-api/src/main/java/io/micronaut/chatbots/telegram/api/ResponseParameters.java
@@ -30,7 +30,7 @@ public class ResponseParameters {
      */
     @Nullable
     @JsonProperty("migrate_to_chat_id")
-    private Integer migrateToChatId;
+    private Long migrateToChatId;
 
     /**
      * In case of exceeding flood control, the number of seconds left to wait before the request can be repeated.
@@ -44,7 +44,7 @@ public class ResponseParameters {
      * @return The group has been migrated to a supergroup with the specified identifier. This number may be greater than 32 bits and some programming languages may have difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or double-precision float type are safe for storing this identifier.
      */
     @Nullable
-    public Integer getMigrateToChatId() {
+    public Long getMigrateToChatId() {
         return migrateToChatId;
     }
 
@@ -52,7 +52,7 @@ public class ResponseParameters {
      *
      * @param migrateToChatId The group has been migrated to a supergroup with the specified identifier. This number may be greater than 32 bits and some programming languages may have difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or double-precision float type are safe for storing this identifier.
      */
-    public void setMigrateToChatId(@Nullable Integer migrateToChatId) {
+    public void setMigrateToChatId(@Nullable Long migrateToChatId) {
         this.migrateToChatId = migrateToChatId;
     }
